refactor(GiniUtilites): Move `Logger` file into a dedicated folder and add code documentation
PP-1107
